CodeUp 1038 solved

+++ CodeUp/basic100/1038.cpp
@@ -5,8 +5,8 @@
 #include <stdio.h>
 
 int main() {
-    int a, b;
-    scanf("%d%d", &a, &b);
-    printf("%d", a+b);
+    long long int a, b;
+    scanf("%lld%lld", &a, &b);
+    printf("%lld", a+b);
     return 0;
 }
